TIP-872: Move category classes to Akeneo\Enrichment
